chore: move typography to another file
