+++ packages/styled/src/index.ts
@@ -18,7 +18,6 @@ import {
 } from './util'
 import { columnSystem } from '@charcoal-ui/foundation'
 import {
-  halfLeading,
   dur,
   GradientDirection,
   notDisabledSelector,
@@ -29,10 +28,10 @@ import {
   Internal,
   internal,
   TRANSITION_DURATION,
-  useHalfLeadingCanceller,
   __DO_NOT_USE_GET_INTERNAL__,
 } from './builders/internal'
 import { createColorCss, createGradientColorCss } from './builders/colors'
+import { createTypographyCss } from './builders/typography'
 export { type Modified, type ModifiedArgumented } from './builders/lib'
 export { default as TokenInjector } from './TokenInjector'
 export {
@@ -268,63 +267,6 @@ const createOutlineColorCss =
       }
     )
   }
-
-const createTypographyCss =
-  <T extends CharcoalAbstractTheme>(theme: T) =>
-  (
-    size: keyof T['typography']['size'],
-    options: {
-      preserveHalfLeading?: boolean
-      monospace?: boolean
-      bold?: boolean
-    } = {}
-  ): Internal => {
-    const {
-      preserveHalfLeading = false,
-      monospace = false,
-      bold = false,
-    } = options
-    const descriptor = theme.typography.size[size]
-    const margin = -halfLeading(descriptor)
-
-    return internal(
-      (context) => ({
-        fontSize: px(descriptor.fontSize),
-        lineHeight: px(descriptor.lineHeight),
-        ...(monospace && {
-          fontFamily: 'monospace',
-        }),
-        ...(bold && {
-          fontWeight: 'bold',
-        }),
-        ...(useHalfLeadingCanceller(context) && {
-          // prevent margin collapsing
-          display: 'flow-root',
-          // cancel half-leading with negative margin
-          '&::before': {
-            ...leadingCancel,
-            marginTop: px(margin),
-          },
-          '&::after': {
-            ...leadingCancel,
-            marginBottom: px(margin),
-          },
-        }),
-      }),
-      !preserveHalfLeading
-        ? {
-            cancelHalfLeadingPx: margin,
-          }
-        : {}
-    )
-  }
-
-const leadingCancel: CSSObject = {
-  display: 'block',
-  width: 0,
-  height: 0,
-  content: `''`,
-}
 
 function spacingProperty(
   property: SpacingProperty,
